refactor: separate sideBar data flow as hook
