+++ resources/assets/js/components/SideBar/index.js
@@ -1,5 +1,5 @@
-import React, { useEffect, useCallback } from 'react';
-import { useLocation, useHistory } from 'react-router-dom';
+import React, { useMemo } from 'react';
+import { useLocation } from 'react-router-dom';
 
 import Drawer from '@material-ui/core/Drawer';
 import List from '@material-ui/core/List';
@@ -10,49 +10,24 @@ import { DrawerContent, useStyle, ListItemText } from './style';
 import { routers, adminRouters } from './routes';
 
 import { useMedia } from '~/utils/index';
+import useSideBarEffect from './useSideBarEffect';
+import useSideBarClick from './useSideBarClick';
 
 const drawerWidth = 110;
 
 export default function SideBar({ open, onClose, onOpen }) {
     const classes = useStyle();
     const device = useMedia();
-
     const location = useLocation();
 
-    const history = useHistory();
-    const arr = location.pathname.startsWith('/admin/')
-        ? adminRouters
-        : routers;
+    const { handleClick, handleToggle } = useSideBarClick({ device, onClose });
+    useSideBarEffect({ device, onClose, onOpen });
 
-    const handleClick = useCallback(
-        (url) => {
-            // 問卷
-            if (url[0] !== '/') {
-                window.open(url, '_blank').focus();
-                return;
-            }
-
-            history.push(url);
-            if (device !== 'PC') onClose();
-        },
-        [history, device, onClose]
+    const arr = useMemo(
+        () =>
+            location.pathname.startsWith('/admin/') ? adminRouters : routers,
+        [location]
     );
-
-    const toggleDrawer = () => (event) => {
-        if (
-            event.type === 'keydown' &&
-            (event.key === 'Tab' || event.key === 'Shift')
-        ) {
-            return;
-        }
-
-        onClose();
-    };
-
-    useEffect(() => {
-        if (device !== 'PC') onClose();
-        else if (device === 'PC') onOpen();
-    }, [device]);
 
     return (
         <Drawer
@@ -60,7 +35,7 @@ export default function SideBar({ open, onClose, onOpen }) {
             className={classes.drawer}
             variant={device === 'PC' ? 'persistent' : 'temporary'}
             open={open}
-            onClose={toggleDrawer('left', false)}
+            onClose={handleToggle('left', false)}
             style={{ width: open ? drawerWidth : 0 }}
             classes={{
                 paper: classes.drawerPaper,
@@ -68,7 +43,7 @@ export default function SideBar({ open, onClose, onOpen }) {
         >
             <DrawerContent>
                 <List component="nav">
-                    {arr.map((item, index) => (
+                    {arr.map((item) => (
                         <ListItem
                             button
                             key={item.text}
