+++ resources/assets/js/components/SideBar/useSideBarEffect.js
@@ -1,0 +1,8 @@
+import { useEffect } from 'react';
+
+export default function useSideBarEffect({ device, onClose, onOpen }) {
+    useEffect(() => {
+        if (device !== 'PC') onClose();
+        else if (device === 'PC') onOpen();
+    }, [device]);
+}
